Feat : 일반 거래처 Model 분리
